requirements.txt
streamlit
langchain
openai
faiss-cpu
tiktoken
unstructured
pypdf
python-dotenv
